[ESP32] Factor out LEDWidget as a component
There were a few copies of LEDWidget, I guess the source-app got
copy-pasted. Refactored the LEDWidget into a component which can be
added to these apps.

Removed LEDWidget from:
 - all-clusters-minimal-app
 - all-clusters-app
 - lighting-app
 - lock-app

and moved it to examples/platform/esp32/led_widget.

This component support LED_TYPE_RMT, LED_TYPE_GPIO, and Virtual LEDs on
M5Stack's display.
