+++ examples/all-clusters-minimal-app/esp32/main/AppTask.cpp
@@ -89,7 +89,7 @@ CHIP_ERROR AppTask::Init()
              (chip_info.features & CHIP_FEATURE_EMB_FLASH) ? "embedded" : "external");
 
     CHIP_ERROR err = CHIP_NO_ERROR;
-    statusLED1.Init(STATUS_LED_GPIO_NUM);
+    statusLED1.Init(static_cast<gpio_num_t>(CONFIG_STATUS_LED_GPIO_NUM));
     // Our second LED doesn't map to any physical LEDs so far, just to virtual
     // "LED"s on devices with screens.
     statusLED2.Init(GPIO_NUM_MAX);
