--- examples/all-clusters-minimal-app/esp32/main/idf_component.yml
@@ -1,2 +1,0 @@
-dependencies:
-    espressif/led_strip: "^3.0.0"
