+++ examples/lighting-app/esp32/sdkconfig_m5stack.defaults
@@ -20,6 +20,7 @@
 CONFIG_IDF_TARGET="esp32"
 CONFIG_IDF_TARGET_ESP32=y
 CONFIG_DEVICE_TYPE_M5STACK=y
+CONFIG_LED_GPIO=5
 
 # Default to 921600 baud when flashing and monitoring device
 CONFIG_ESPTOOLPY_BAUD_921600B=y
