--- examples/lock-app/esp32/main/include/LEDWidget.h
@@ -1,45 +1,0 @@
-/*
- *
- *    Copyright (c) 2020 Project CHIP Authors
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-#pragma once
-#include "driver/gpio.h"
-#include <stdint.h>
-
-#ifndef LED_WIDGET_H
-#define LED_WIDGET_H
-
-class LEDWidget
-{
-public:
-    void Init(gpio_num_t ledNum);
-    void Set(bool state);
-    void Invert(void);
-    void Blink(uint32_t changeRateMS);
-    void Blink(uint32_t onTimeMS, uint32_t offTimeMS);
-    void Animate();
-
-private:
-    uint64_t mLastChangeTimeMS;
-    uint32_t mBlinkOnTimeMS;
-    uint32_t mBlinkOffTimeMS;
-    gpio_num_t mGPIONum;
-    bool mState;
-
-    void DoSet(bool state);
-};
-
-#endif // LED_WIDGET_H
