+++ examples/platform/esp32/led_widget/CMakeLists.txt
@@ -1,0 +1,6 @@
+idf_component_register(
+    SRCS "LEDWidget.cpp"
+    INCLUDE_DIRS "include"
+    REQUIRES esp_driver_gpio
+    PRIV_REQUIRES driver esp_timer
+)
