+++ examples/platform/esp32/led_widget/idf_component.yml
